Revert "refactor common types into its own util class"
This reverts commit 3dbadc6e6d627d0336393b9c4c0914cf7409722d.
