--- src/urllib3/util/typing.py
@@ -1,3 +1,0 @@
-from typing import IO, Any, Iterable, Union
-
-_TYPE_BODY = Union[bytes, IO[Any], Iterable[bytes], str]
